faft: Change the search pattern of ECBin DTB
The pattern of ECBin DTB in ToT is changed from "blob boot,dtb,ecbin" to
"blob boot,dtb-rwa,ecbin". This CL is to change the pattern to match both.

BUG=chrome-os-partner:16680
TEST=Ran the firmware_UpdateECBin test on Link firmware 2695.1.136 and 3558.
It passed after skipping the ecfw_act check which is a work-around of some
other issue.

Change-Id: I09e9cb32ed856cf1174c9230455ca9a517eccfaa
Reviewed-on: https://gerrit.chromium.org/gerrit/41284
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>