client/site_tests: add power_VideoDetector test
This test plays video with reduced powerd timeouts and makes sure that
the video is being detected by power manager and the backlight doesn't
get dimmed.

The "video" is an html file with webkit code that updates the color of
the page.

Also moves some of the common code across a few power autotests into
common libraries:
- Functions to adjust backlight using powerd
- Function to restart an upstart job
- Function to overwrite/backup powerd prefs

BUG=chromium-os:32371,chromium-os:36382
TEST=power_VideoDetector passes

Change-Id: I01c8fb63368778aa6910191a9e32412bb502f691
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/41414
Reviewed-by: Daniel Erat <derat@chromium.org>