[autotest] Disable 'F' in pylint.
Pylints import directory resolution model isn't the same
as the python interpreters, so at times it believes it
cannot import a module when the interpreter can. This
failure is falky and depends on the path ordering/environment;
it doesn't repro on my machine. Disabling the F mode for now.

BUG=chromium-os:37948
TEST=Made sure it worked in the environment it failed in.

Change-Id: I3a5f91a066d9e45d1db82135213675c109c3420a
Reviewed-on: https://gerrit.chromium.org/gerrit/41429
Commit-Queue: Prashanth Balasubramanian <beeps@chromium.org>
Reviewed-by: Prashanth Balasubramanian <beeps@chromium.org>
Tested-by: Prashanth Balasubramanian <beeps@chromium.org>