Remove servo requirement from full_release_test and schedule remote jobs.
This CL is composed of two components:

1) Remove the hardcoding of localhost to use the afe defined in the autotest
config. This allows for either local/remote but defaults to using cautotest.
2) Removes the servo requirement. If servo is not set, the full payload for
the src image is sent as opposed to the test image for the src. Also, servo
args are not used.

This CL also does some small cleanup strings and instead of reporting the
job_id, gives the full URL to the job on any debug level.

BUG=chromium-os:37734, chromium-os:37735
TEST=Ran it against cautotest with --npo, --nmo, --fsi. run_pylint and friends.

Change-Id: If5bdd73ec1242fc863b81cfa57fc6e0a9ab1024b
Reviewed-on: https://gerrit.chromium.org/gerrit/41502
Tested-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Gilad Arnold <garnold@chromium.org>
Commit-Queue: Chris Sosa <sosa@chromium.org>