Add network_ONCServer test to night runs
Add the network_ONCServer test to the nightly runs.
they require the use of the wifi cells, for now, we want
to isolate the usage to the one located in oysterbay.  Due to massive
updates to the test, we only want this running for >R26 devices.

BUG=None
TEST=Make sure the results start showing up for network_ONCServer.

Change-Id: I277d4ffba1df7e0a7af84c1201899a7df3d24e5b
Reviewed-on: https://gerrit.chromium.org/gerrit/41594
Commit-Queue: Stanley Wong <stanleyw@chromium.org>
Reviewed-by: Stanley Wong <stanleyw@chromium.org>
Tested-by: Stanley Wong <stanleyw@chromium.org>