Fix EC UART interactions
When FAFT is running on a setup where the servo board is controlled by
a Beaglebone board, many EC tests fail because the expected output
from the EC console is not received in the allotted time.

Actually, there is no need to limit the time when the test is waiting
for the response: if the expected response comes, the wait stops right
there, so the timeout does not matter. If the response does not show
up, a few extra seconds will not make things much worse.

This change deprecates the timeout parameter, all uses of
send_command_get_output() will run with the same previously set
timeout. By default the timeout is set to 3s in the servod package
(https://gerrit.chromium.org/gerrit/41568).

If some tests need a longer timeout, or there is a point in specifying
a short timeout, it still could be done using the `ec_uart_timeout'
control.

BUG=chromium-os:22475
TEST=manual
  . successfully ran the following tests in two setups, with local and
    remote servos:

    firmware_ECAdc
    firmware_ECBattery
    firmware_ECBootTime
    firmware_ECCharging
    firmware_ECKeyboard
    firmware_ECPeci
    firmware_ECPowerG3
    firmware_ECSharedMem
    firmware_ECThermal
    firmware_ECUsbPorts
    firmware_ECWakeSource
    firmware_ECWatchdog
    firmware_ECWriteProtect
    firmware_FAFTSetup

Change-Id: I0da2f9a0924574846a6c7650e996576a7d2b76aa
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/41618
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>