Touch FW test: Change "Sample" rate to "Report"
No functional changes here, just that the rate at which the events are
reported back to the kernel by the touchpad FW was referred to as the
"sample" rate, which is a bit of a misnomer since in other places the
"sample" rate is defined as the rate at which the fw scans the finger
locations and how often it reports back to the kernel is called the
"report" rate.  This is fine for people who are aware of the difference,
but it seems like it would be less prone to confusion if it was just
changed to match.

BUG=chromium-os:37494
TEST=manually ran the test suite and made sure everything still ran and
I hadn't missed any of the "sample" rates when changing them over.

Change-Id: I0bc0717e346334d31d20f37b79981333188a7c87
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/41764
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>