Separate FAFT tests to improve run time
The tests are capable to moving the DUT into the required state (dev
or normal), but frequent state change takes a lot of time. It is
better to run tests in series by the initial state (dev/control). The
EC tests are in their own category. This will also allow to run tests
concurrently on multiple DUTs.

There are a few tests which have 'faft' in the SUITE = section, but
are not classified as either dev, normal or ec - they will be assigned
to one of the categories in a separate change.

BUG=chromium-os:38129
TEST=manual
   $ ./site_utils/suite_scheduler/suite_scheduler.py --sanity

Change-Id: Ibab67f1648e4a9259de259f3a41dbe329eb3475f
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/41807
Reviewed-by: Scott Zawalski <scottz@chromium.org>