Fix a small issue with firmware test email.
We had dup firmware filters. The first firmware filter was
blocking the second. This should enable subsequent faft email.
However an ongoing issue will remain (http://crosbug.com/31934).
This might help a little - it won't hurt.

The emailer (the -m option on run_generate) has an override
built in for testing. If the running user is not the system
account (chromeos-test), at the very end email is actually
sent to the user running the program. So, this TEST= activity
sends expected emails to me and I can verify from:, to:, cc:,
content.

BUG=chromium-os:38203
TEST=run_generate -j 1500 -m -v

To Test this, run_generate --help shows options.
For emailer testing the following are recommended:
-j Sets the data query to look back 1500 jobs
-m Sets the dashboard generator into emailer mode (vs table build mode)
-v Enabled debug spew.

After doing this, when emails are intended, the test emailer will log a
line with the to: from: details that is viewable.  Then, the actual
message to: is overridden and sent to the logged in user.

Change-Id: I65ac79433cd6785b0831c80ea50bb99023004a33
Reviewed-on: https://gerrit.chromium.org/gerrit/41946
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Commit-Queue: Mike Truty <truty@chromium.org>
Tested-by: Mike Truty <truty@chromium.org>