Telemetry on Autotest: Telemetry runner code and simple test
This change adds the code to have the drones and server side tests
able to execute the telemetry code installed on the devserver.

The telemetry runner will determine the correct devserver for the
DUT's specific build, tell the devserver to setup telemetry and given
a benchmark and page_set it will execute a telemetry benchmark and
return a TelemetryResultObject to represent the results of telemetry's
execution.

The stdio of the telemetry output is also parsed and used to generate
perf key-value pairs to be uploaded into the autotest results db.

BUG=chromium-os:37412, chromium-os:38997
TEST=tested the telemetry side with my local AFE and devserver and
tested the pyauto proxy portion in the actual lab.

Change-Id: Ia8da3f991b498e1ee63fbfffbfaaa628e3785bf3
Reviewed-on: https://gerrit.chromium.org/gerrit/43679
Commit-Queue: Simran Basi <sbasi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>