Include dts subdirectory in the set of extracted files
ARM platforms re-imaging requires that the device tree files (now
included in subdirectory ./dts in the firmware archive) are available
to the programmer along with the actual firmware image. This change
includes the entire contents of the dts/ subdirectory into the set of
extracted files.

To maintain the ability to reimage X86 platforms using tarballs
generated before device trees were included, extracting of the device
trees is done in a separate invocation allowing the tar command to
fail if the device trees are not present.

Added DUT reset after reimaging is done to make sure the new firmware
kicks in.

And, to make testing easier, this change allows use of yet another
type of URL (file://) for specifying the location of the firmware
archive.

BUG=chrome-os-partner:15610
TEST=manual

 . run the following commands:
  $ emerge-link autotest
  $ gsutil cp gs://chromeos-releases/canary-channel/link/3636.0.0/ChromeOS-firmware-R26-3636.0.0-link.tar.bz2 /tmp/link.tar.bz2
  $ run_remote_tests.sh --use_emerged --board=link --servo \
    --remote=192.168.1.4 --allow_offline_remote fwupdate --args \
    'fwurl=file:///tmp/link.tar.bz2 board=link servo_host=192.168.1.102'

Change-Id: I730284c0959bea7b2486c6b1376a960e9e658597
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/42342
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>