Reenable power_SuspendStress in lab suites and improve suspend timings
My last attempt to get this running in the lab was a little premature and
ended up leaving too many test machines stuck in suspend. This patch
contains trial-and-error-derived timing values for every ChromeOS device
currently in the lab, and should finally make this run without causing
interruptions. The bare and disk tests get reactivated to run in their
respective suites... but I will leave the video test out for now since
it triggers known hard freezes on several platforms.

BUG=chromium-os:35901
TEST=None

Change-Id: Icaea3d3497677095cfedea6c65fc8e1eb81f6d72
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/42431
Reviewed-by: Mark Koudritsky <kamrik@chromium.org>