Fix read_all_logs regression on rotated logs
In CL:41677, I "simplified" the start line detection in
cros_logging.read_all_logs(). Unfortunately, I forgot the case of
multiple log files, in which my code starts discarding lines again for
every new file and ends up reading over the end of the file.

This patch fixes that issue by returning to the old way this was
handled.

BUG=None
TEST=Run power_Resume on a machine with 2+ /var/log/messages.* files.

Change-Id: I5a195614c4820c621547dc9290a257dfceb593d4
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/42354
Reviewed-by: Mark Koudritsky <kamrik@chromium.org>