Factory: Allow the VPD test to fall back to manual selection.
If a factory shopfloor server has no data for a device, allow
the operator to manually enter the information. If the shopfloor
server throws an error, the VPD test will still fail.

BUG=chrome-os-partner:17302
TEST=Run the factory flow with a shopfloor server that returns no
data for a GetVPD request instead of throwing an error. Verify
that the operator is then allowed to manually select the VPD options.

Change-Id: I1744c2b2b7952373f5f3c44d027e8b6641bdddd8
Reviewed-on: https://gerrit.chromium.org/gerrit/42430
Tested-by: Dave Parker <dparker@chromium.org>
Reviewed-by: Chinyue Chen <chinyue@chromium.org>
Commit-Queue: Dave Parker <dparker@chromium.org>