Move ServiceStopper out of upstart.py
ServiceStopper needs to import utils to keep it from breaking some of the
autotests that it is used in. This is not allowed in upstart.py per the comment
at the head of the file. This CL moves ServiceStopper to its own file
service_stopper.py and adjusts the affected tests appropriately.

BUG=chromium-os:38544
BUG=chromium-os:38561
TEST=Build and run the affected autotests:
        - graphics_SyncControl
        - hardware_Backlight
        - power_Backlight
        - power_BatteryCharge
        - power_Consumption
        - power_Idle

Change-Id: Icde86185ffd66d399bac10debe3dee7e8d6cf45b
Reviewed-on: https://gerrit.chromium.org/gerrit/42536
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>