Override devserver_up's default timeout to return immediately in au test.
Since we control this devserver, we want to control how long to wait for
the devserver to be up. This CL slightly refactors the devserver api
to allow callers to override the retry.retry timeout in some rpc's.

Also fixed the pylintrc to skip **kwargs, *args, and cls.

BUG=None
TEST=Pylint so far + unittests

Change-Id: Ie454e81be8940bdc7b8530bd682dab80910b8b83
Reviewed-on: https://gerrit.chromium.org/gerrit/43480
Tested-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Brian Harring <ferringb@chromium.org>
Reviewed-by: Prashanth Balasubramanian <beeps@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Chris Sosa <sosa@chromium.org>