Remove target_hostname from Servo objects.
The function currently found in server.cros.servo needs to be able
to move to a class in hdctools.  The presence of a 'target hostname'
attribute isn't really appropriate for hdctools, and moreover serves
no useful purpose in autotest code, either.  Remove the attribute,
and fix the references to do the right thing.

BUG=chromium-os:36973
TEST=run firmware_FaftSetup on a DUT in the cros lab

Change-Id: I18e0ad94f04466294007fb34a0f9e3d6ad5e0eae
Reviewed-on: https://gerrit.chromium.org/gerrit/43075
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Vic Yang <victoryang@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>