Autotest now logs the hwid as a test keyval.
As part of investigating performance test results, it's helpful to have
information about the physical device on which the tests ran.  This CL
adds the "hwid" to the set of test keyvals outputted by an autotest run.
From the hwid, we can look up detailed information about a device.

This CL also adds some docstrings to existing code in order to get
presubmit checks to pass.

BUG=chromium-os:26401
TEST=Verified the hwid appears in the test "keyval" file when running a
sample test on a local device.

Change-Id: I886762900171234bce3897ed74f0a14d69b3a928
Reviewed-on: https://gerrit.chromium.org/gerrit/43078
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Queue: Dennis Jeffrey <dennisjeffrey@chromium.org>
Tested-by: Dennis Jeffrey <dennisjeffrey@chromium.org>