EagleEye perf graphs now support autotests whose names have changed.
The JSON file specifying which performance graphs to plot has now been
restructured: the file is now a list of objects representing each test
to plot.  A test has a name, an optional list of "old_test_names" (in
case the name of the test has changed over time), and a list of graphs
to display that are associated with the test.

The scripts to extract perf values and generate perf graphs have been
adjusted to account for the changed JSON file format.  In the event that
a test is associated with one or more "old_test_names", then any perf values
from an old test name are processed and treated as if they are simply
associated with the new test name.  So a test that used to be called "X",
but is now called "Y", will show up in the graphs as test "Y", but will
implicitly include all results from both "X" and "Y" taken together.

BUG=chromium-os:38150
TEST=Verified locally that autotest desktopui_PyAutoPerf (which used to
be called desktopui_PyAutoPerfTests in the autotest database) shows up
properly in the perf graphs and contains the perf results from both
the old and new autotest names.

Change-Id: I5b41e176b5fa72f0156984b067cd14dc56f3d73b
Reviewed-on: https://gerrit.chromium.org/gerrit/43238
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Dennis Jeffrey <dennisjeffrey@chromium.org>
Tested-by: Dennis Jeffrey <dennisjeffrey@chromium.org>