Prevent autotest dbus-suspend attempt when no user is logged into device.
When no user is currently logged into a device and an attempt is made to
suspend the device using dbus, the device actually shuts down instead of
suspending.  This effectively kills the device until someone goes and boots
it back up again.

This change identifies when an attempt is made to suspend a device using
dbus when no user is currently logged in.  In that situation, an exception
is raised to alert the user and prevent the suspend from being attempted.

Also made several other changes to fix presubmit warnings.

BUG=chromium-os:38140
TEST=Verified that when I run autotest power_Resume and force it to
suspend using dbus, then:
1) Before this CL, when no user is currently logged in, then running the
autotest causes the device to shut down.
2) After this CL, when no user is currently logged in, then running the
autotest causes the test to fail with an exception (and the device remains
running).
3) After this CL, when a user is logged in, then running the autotest causes
the test to pass (and the device remains running).

Change-Id: Icd50c86d7f8be681bcee7dd01e66b5667d86b9e5
Reviewed-on: https://gerrit.chromium.org/gerrit/43202
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Queue: Dennis Jeffrey <dennisjeffrey@chromium.org>
Tested-by: Dennis Jeffrey <dennisjeffrey@chromium.org>