[autotest] Fix unittest -- chromeos_test/test_config.py
Modified a file path to default config file in test_config.py so that
unittest_suite will work when invoked from the root directory of the
autotest repo.
Added an entry to a list in the assertion for testValidConfig to get
test to pass.

BUG=chromium-os:37989
TEST=Unit tests for this module now pass

Change-Id: I17f2aae352eb89acfb24c3ed54f133c8ed0940f4
Reviewed-on: https://gerrit.chromium.org/gerrit/43209
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>