[autotest] support file path in blacklist in unittest_suite.py
Code is changed to check if a file is in the blacklist with both
file name and path. For example, 'mirror/trigger_unittest.py' in
REQUIRES_AUTH set will skip the unittest in mirror folder but not
the one in any other folder, while 'trigger_unittest.py' will skip
any test with that file name in any folder.

BUG=chromium-os:28739
TEST=run unittest_suite.py locally to verify files in blacklist
are skipped.

Change-Id: I3f5ab5fab83785f9d55241661d37919390753ed0
Reviewed-on: https://gerrit.chromium.org/gerrit/43218
Tested-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Dan Shi <dshi@chromium.org>