[autotest] Fix unittest -- mp_thread_pool
The mp_thread_pool unittest is currently failing. Looks like this is
because MultiProcWorkPool().__init__(...) is not correctly handling
procs=0. The docstring indicates that procs=0 should result in no
processing being created, but instead the function attempts to call
mp.Pool(processes=0), resulting in an exception.

Bug crosbug.com/38902 has been filed, and for now this commit disables
the unit tests for this module.

BUG=chromium-os:37989
TEST=Disabled test no longer runs as part of utils/unittest_suite.py

Change-Id: Iebe2e89a84f6ee49554aca0cd69d4be461146dc8
Reviewed-on: https://gerrit.chromium.org/gerrit/43234
Commit-Queue: Aviv Keshet <akeshet@chromium.org>
Reviewed-by: Aviv Keshet <akeshet@chromium.org>
Tested-by: Aviv Keshet <akeshet@chromium.org>