Remove host argument from Servo.install_recovery_image().
Servo.install_recovery_image() accepted a host argument; that
argument wasn't appropriate for this file and the plan to move its
contents to hdctools.  The argument was only supplied for the
specific case of platform_InstallRecoveryImage, so the code moved
there instead.

The code that used the host argument also didn't work well for
recovery images, because it relied on the DUT to respond to ssh
after recovery, which normally wouldn't happen.  This change updates
the test to use ping in place of ssh.

BUG=chromium-os:36973
TEST=run platform_InstallRecoveryImage

Change-Id: I82dc346b3b5c96d6b818a249588a07d9f021985c
Reviewed-on: https://gerrit.chromium.org/gerrit/43310
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>