Small refactoring / cleanup to allow autotest to continue working with devserver API.
Not much to this CL -- lots of small changes which make it continue
to work, and in some places work much better/faster.

1) Add a method to use the stage RPC directly (not cleaning up
trigger/finish_download in this CL
2) I removed the implicit download of the firmware tarball from the
spatial locality map -- it always errors out and pretty much hogs our
devserver logs. Instead I stage the firmware tarball when we need it
and fail that call as necessary. This is much more correct than what we have
now.
3) Removed test image RPC from dev_server and changed its one callspot to use
(1)
4) Fixed the bug in symoblicating crashes -- before there was a race to whoever
last staged symbols got to symoblicate crashes -- not anymore as we store the symbols
per build in a subdir for that build.
5) Cleaned up autoupdate test to stage only what it needs.

Consult I201dcdaa8a14024247 for information about artifact naming.

BUG=chromium-os:38427
TEST=In progress + unittests
CQ-DEPEND=I201dcdaa8a14024247

Change-Id: I64e35f3bdeac9e0cd5e7589df913c14ce6691f7a
Reviewed-on: https://gerrit.chromium.org/gerrit/43350
Reviewed-by: Chris Sosa <sosa@chromium.org>
Tested-by: Chris Sosa <sosa@chromium.org>