Modify fwupdate to run smoothly even when the DUT is down
We want to be able to run fwupdate even when DUT is down and can not
respond to attempts to establish an ssh connection. There is no need
to create the host object in this case, so we create a servo object
directly instead of borrowing it from the host object.

BUG=chrome-os-partner:15610
TEST=manual
   . corrupt DUT firmware so that it does not come out of reset
   . observe the following test succeed:

   $ run_remote_tests.sh --servo --allow_offline_remote --use_emerged --board=daisy \
     --remote=192.168.1.116 fwupdate --args \
    'fwurl=gs://chromeos-releases/canary-channel/daisy/3721.0.0/ChromeOS-firmware-R27-3721.0.0-daisy.tar.bz2 \
    board=snow'

Change-Id: I661e992b23423525d13e932e0e5ceb2b3b0c1ec8
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/43501
Reviewed-by: Alex Miller <milleral@chromium.org>