autotest: WiFi: Correct default SSID of 'disconnect' step function.
In below wifi autotests, they have a 'disconnect' step without specifying SSID.
    network_WiFiManager/092_ProfileCredentials
    network_WiFiManager/094_ProfileDelete

There is a default value of SSID in 'connect' and 'disconnect' steps if the
test case does not specify SSID.
The default SSID in function 'disconnect' is self.__get_defssid().
However, the default SSID in function 'connect' is self.wifi.get_ssid()
(self.wifi.get_ssid() is the current SSID of hostapd).

The mistmatch of SSID causes the failure of 'disconnct' step.
Change default SSID of 'disconnect' to be the same as that of 'connect'.

BUG=chromium-os:35374
TEST=Run wifi autotest: network_WiFiManager/092_ProfileCredentials
                        network_WiFiManager/094_ProfileDelete

Change-Id: I4c392e2b4f104c4b027d60935a7406533a548fb5
Reviewed-on: https://gerrit.chromium.org/gerrit/43505
Tested-by: Bowgo Tsai <bowgotsai@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Bowgo Tsai <bowgotsai@chromium.org>