firmware_TouchpadMTB: compute the final weighted average
Currently, we provide the test scores based on gestures and
also on validators between devices/firmwares. For example,

Test Summary (by gesture)   : cyapa_fw_1.0  cyapa_fw_1.1
--------------------------------------------------------
one_finger_tracking
  CountTrackingIDValidator  : 1.00          0.87
  LinearityBothEndsValidator: 0.98          0.92
   ...
two_finger_tracking
...

Test Summary (by validator) : cyapa_fw_1.0  cyapa_fw_1.1
--------------------------------------------------------
CountPacketsValidator       : 0.90          0.85
CountTrackingIDValidator    : 1.00          0.89
...

It is not straightforward to tell which device/firmware is better.
In this patch, we compute the weighted average of the scores of
all validators. Every validator is assigned a relative weight to
indicate its importance. In this way, it is easier to tell which
device/firmware is better based on a final single score. For example,

Test Summary (final weighted average)
-------------------------------------
cyapa_fw_1.0: 0.902
cyapa_fw_1.1: 0.913

BUG=chromium-os:35024
TEST=Run the firmware_summary module either on your host or
on a chromebook machine.
$ python firmware_summary.py path/to/your/logs
It lists the final weighted average of the test scores.

Change-Id: I3f9489851c51fb2b300956517633d641c0312e5e
Reviewed-on: https://gerrit.chromium.org/gerrit/43713
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>