firmware_TouchpadMTB: use relative path for bin/input
Sometimes we may want to execute the firmware_summary module on a
host instead of a chromebook. In this patch, we append the relative
path of bin/input so that the path could be used in the environment
of either a host or a chromebook.

BUG=chromium-os:35024
TEST=Run the test command in the chroot of your host:
(cr) $ cd ~/trunk/src/third_party/autotest/files/client/site_tests/\
firmware_TouchpadMTB
(cr) $ python firmware_summary.py path/to/your/logs
The firmware_summary module should be executed normally.

Change-Id: If16c06883d33332a0a4b179d66c9cd70c7217604
Reviewed-on: https://gerrit.chromium.org/gerrit/43711
Reviewed-by: Charlie Mooney <charliemooney@chromium.org>
Commit-Queue: Joseph Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Joseph Shyh-In Hwang <josephsih@chromium.org>