Remove login_CryptohomeMounted from smoke temporarily.
The login_CryptohomeMounted test has failed over 86 times over the past 8
days. There is nothing wrong with this test, per se, but it happens to be very
good at reproducing bug 177357. For now, we are disabling the test to give the
Chrome team time to fix bug 177357.

As soon as bug 177357 is fixed, we should revert this CL. This should ideally
happen within one week.

BUG=chromium:177357
TEST=None

Change-Id: I4da8e087e341b8c7e0edb1ecf2dfe0bbb3c081e5
Reviewed-on: https://gerrit.chromium.org/gerrit/43671
Reviewed-by: Zelidrag Hornung <zelidrag@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>