Switch full_release_test to use a class to pass args.
Was adding a flag and it took almost 100 lines to propogate to
find_payload_uri which was a lot of work to test. Instead, I have
refactored to put all these methods into a class such that
adding a new option to get propagted is < 10 lines of code making
the code easier to maintain/work on.

BUG=chromium-os:37575
TEST=Ran --npo, --nmo, --specific.

Change-Id: I4046eae0788f75fd440b34078b6056e0f7f4f18c
Reviewed-on: https://gerrit.chromium.org/gerrit/43839
Tested-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Gilad Arnold <garnold@chromium.org>
Commit-Queue: Chris Sosa <sosa@chromium.org>