[Autotest] Update the Autotest installer code to avoid reinstalling the same version
When a try job is started to install a new image in DUT, the DUT's version
is checked first. If the version is the same as the new image, machine_install
function in SiteHost first try to use stateful update to clean up the machine.
This is implemented by adding a new argument, update_root, in function
run_update in autoupdater. When update_root is False, autoupdater will only
do a stateful update, compared to updating kernel as well when update_root
is True. The main purpose for this change is to enhance the performance of
the try job on re-imaging.

BUG=chromium-os:31239
TEST=run_sutie against lumpy with smoke suite, try the command with same/
different version compared to what's running in DUT.

Change-Id: Ie358e7a5282f1fadefc26e78bbe4b38cd8e8fb8a
Reviewed-on: https://gerrit.chromium.org/gerrit/43851
Tested-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Commit-Queue: Dan Shi <dshi@chromium.org>