Renabled the ecryptfs tests
With the fixing of the mount command for ecryptfs, we can now start
running the ecryptfs tests.
The test will add about 9 minutes for x86 tragets and 30 minuets for
snow.

Had to update comments to get through new upload checks.

Added control.daily so tests would run as part of daily tests.
However, did not include disk and ecyptfs tests because they
are failing in certain cases.

I've failed to defects to for the problems I've found:
http://crosbug.com/39627 related to interactions with swap
http://crosbug.com/39629 system crashes if tests are run twice

BUG=chromium-os:39259
TEST=./run_remote_tests.sh --use_emerged --board=$B --remote=$IP client/site_tests/kernel_fs_Punybench/control.daily

Change-Id: I8b7b563ecd1c0e1fd8891b39625df9cf2f52affd
Reviewed-on: https://gerrit.chromium.org/gerrit/44252
Tested-by: Paul Taysom <taysom@chromium.org>
Reviewed-by: Mike Truty <truty@chromium.org>
Commit-Queue: Paul Taysom <taysom@chromium.org>