autotest: WiFi: Collect AP capabilities and requirements
Collect a set of capabilities of the AP (whether it can do
5GHz, whether it can support multiple simultaneous APs and
whether it can do multiple APs on the same frequency band).
While scanning through the steps, get a list of both router
and client capabilities required.

This will allow us to error-out early if the AP is not
configured in a manner that will allow the test to proced.
For example, it will confirm whether it's possible to run
an multi-AP test.  Both this new error and the old client
capability error should raise TEST_NA instead of completing
silently.

BUG=chromium-os:38227
TEST=Run network_WiFiManager/093_ProfilePopPush, a test that
requires both 5ghz and multi_ap and confirm that this appears
in the step requirements.  Ensure that both correctly
configured stumpy cells and netgear cells can run this test
successfully, and that if one interface is removed from a
stumpy cell, the test fails with TEST_NA.  Also confirm that
a correctly running stumpy cell also reports multi_ap_same_band.

Change-Id: I3f7ce1df2c749fb9aaadbc2195f30464be00d72a
Reviewed-on: https://gerrit.chromium.org/gerrit/44270
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>