Touch FW test suite Keyboard detection for Snow
There was a problem with how the keyboard device was detected on Snow
because it isn't registed as the same type as on x86 so the regex
doesn't match.  This adds the additional keyboard option so it will
match on the ARM version of cros as well.

BUG=chromium-os:39360
TEST=Installed on a snow and ran, it was able to connect to the keyboard
without incident now

Change-Id: I583bda1baed67e590c5048076ba4cdff27fd2b81
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/44282
Reviewed-by: Joseph Shyh-In Hwang <josephsih@chromium.org>