`Convert 'Timed out waiting to revert DNS' into a TestWarn
This error condition doesn't actually mean that the test in question
failed to complete.  It means that, while trying to verify that the
device state was successfully back to what it was before the test ran,
we had a problem.  Things might be fine.  Therefore, if subsequent tests
can succeed...things are good to go.  That said, if subsequent tests fail,
it'll be helpful to have this message reported somewhere visible.

To do this, we raise a TestWarn exception, which our tooling treats
as a non-fatal failure.

BUG=chromium-os:39147
TEST=manual
STATUS=Fixed

Change-Id: I251482d871f1824be7f85f109b47f525f298fa1c
Reviewed-on: https://gerrit.chromium.org/gerrit/44293
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>